requirements.txt
absl-py==2.2.1
accelerate==0.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.2
astunparse==1.6.3
attrs==25.3.0
banal==1.0.6
beautifulsoup4==4.13.3
bitsandbytes==0.41.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.10
contourpy==1.3.1
cycler==0.12.1
dataset==1.6.2
datasets==3.5.0
dill==0.3.8
evaluate==0.4.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
googlesearch-python==1.3.0
greenlet==3.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
onnxruntime-directml==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optimum==1.24.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
pywin32==309
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==1.4.54
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tokenizer==3.4.5
tokenizers==0.14.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.35.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.92
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
